chore: parse json data in api-provider-environment
