+++ libs/dapp-config/src/bridges/bridge-config.ts
@@ -3,17 +3,10 @@ import { CurrencyRole } from '@webb-tools/dapp-types';
 import { ApiConfig } from '../api-config';
 import { CurrencyConfig } from '../currencies';
 
-let bridgeConfigByAsset: ApiConfig['bridgeByAsset'];
-
 export const getBridgeConfigByAsset = (
   currencies: Record<number, CurrencyConfig>,
   anchors: ApiConfig['anchors']
 ): ApiConfig['bridgeByAsset'] => {
-  // If the bridge config is already calculated, return it
-  if (bridgeConfigByAsset) {
-    return bridgeConfigByAsset;
-  }
-
   const bridgeByAsset: ApiConfig['bridgeByAsset'] = {};
 
   const fungibleCurrencies = Object.values(currencies).filter(
@@ -31,8 +24,6 @@ export const getBridgeConfigByAsset = (
       anchors: anchors[currency.id],
     };
   });
-
-  bridgeConfigByAsset = bridgeByAsset;
 
   return bridgeByAsset;
 };
