+++ libs/dapp-config/src/constants.ts
@@ -1,2 +1,2 @@
 export const ON_CHAIN_CONFIG_PATH =
-  'libs/dapp-config/src/generated/on-chain-config.json';
+  'libs/api-provider-environment/src/generated/on-chain-config.json';
