+++ libs/dapp-config/src/types.ts
@@ -1,5 +1,7 @@
 // File contains shared types for dapp-config
 
+import { ICurrency } from './on-chain-config';
+
 /**
  * The type of the environment the dapp is running in
  */
@@ -13,3 +15,30 @@ export const isAppEnvironmentType = (env: string): env is AppEnvironment => {
     env === 'production'
   );
 };
+
+export interface AnchorMetadata {
+  /**
+   * The address of the anchor
+   */
+  address: string;
+
+  /**
+   * The fungible currency of the anchor
+   */
+  fungibleCurrency: ICurrency;
+
+  /**
+   * The wrappable currencies of the anchor (excluding native currency)
+   */
+  wrappableCurrencies: ICurrency[];
+
+  /**
+   * Boolean indicating whether native currency is allowed
+   */
+  isNativeAllowed?: boolean;
+
+  /**
+   * Record of linkable typed chain -> anchor address
+   */
+  linkableAnchor: Record<string, string>;
+}
