+++ libs/dapp-config/src/utils/index.ts
@@ -1,2 +1,3 @@
 export { default as addCurrencyToConfig } from './addCurrencyToConfig';
+export { default as parseOnChainData } from './parseOnChainData';
 export * from './getNativeCurrencyFromConfig';
